refactor(e2e): organize spec files into feature-based folders per #9671
