+++ frontend/tests/e2e/apply/forms/failure-path/specs/failure-path-cd511.spec.ts
@@ -14,7 +14,7 @@ import {
   CD511_FORM_MATCHER,
   CD511_REQUIRED_FIELD_ERRORS,
 } from "tests/e2e/apply/fixtures/cd511-field-definitions";
-import { getOpportunityId } from "tests/e2e/get-opportunityId-utils";
+import { getOpportunityId } from "tests/e2e/utils/get-opportunityId-utils";
 import playwrightEnv from "tests/e2e/playwright-env";
 import { VALID_TAGS } from "tests/e2e/tags";
 import { authenticateE2eUser } from "tests/e2e/utils/authenticate-e2e-user-utils";
