+++ frontend/tests/e2e/apply/forms/failure-path/specs/failure-path-sf424.spec.ts
@@ -14,7 +14,7 @@ import {
   SF424_FORM_MATCHER,
   SF424_REQUIRED_FIELD_ERRORS,
 } from "tests/e2e/apply/fixtures/sf424-field-definitions";
-import { getOpportunityId } from "tests/e2e/get-opportunityId-utils";
+import { getOpportunityId } from "tests/e2e/utils/get-opportunityId-utils";
 import playwrightEnv from "tests/e2e/playwright-env";
 import { VALID_TAGS } from "tests/e2e/tags";
 import { authenticateE2eUser } from "tests/e2e/utils/authenticate-e2e-user-utils";
