+++ frontend/tests/e2e/apply/forms/failure-path/specs/failure-path-sf424d.spec.ts
@@ -14,7 +14,7 @@ import {
   SF424D_FORM_MATCHER,
   SF424D_REQUIRED_FIELD_ERRORS,
 } from "tests/e2e/apply/fixtures/sf424d-field-definitions";
-import { getOpportunityId } from "tests/e2e/get-opportunityId-utils";
+import { getOpportunityId } from "tests/e2e/utils/get-opportunityId-utils";
 import playwrightEnv from "tests/e2e/playwright-env";
 import { VALID_TAGS } from "tests/e2e/tags";
 import { authenticateE2eUser } from "tests/e2e/utils/authenticate-e2e-user-utils";
