+++ frontend/tests/e2e/apply/forms/happy-path/specs/happy-path-epa4700-4.spec.ts
@@ -12,7 +12,7 @@ import {
 } from "@playwright/test";
 import { EPA4700_4_FORM_CONFIG } from "tests/e2e/apply/fixtures/epa4700-4-field-definitions";
 import { epa4700_4HappyPathTestData } from "tests/e2e/apply/fixtures/epa4700-4-fill-data";
-import { getOpportunityId } from "tests/e2e/get-opportunityId-utils";
+import { getOpportunityId } from "tests/e2e/utils/get-opportunityId-utils";
 import playwrightEnv from "tests/e2e/playwright-env";
 import { VALID_TAGS } from "tests/e2e/tags";
 import { authenticateE2eUser } from "tests/e2e/utils/authenticate-e2e-user-utils";
