+++ frontend/tests/e2e/apply/forms/happy-path/specs/happy-path-sf424d.spec.ts
@@ -11,7 +11,7 @@ import {
 } from "@playwright/test";
 import { SF424D_FORM_CONFIG } from "tests/e2e/apply/fixtures/sf424d-field-definitions";
 import { sf424DHappyPathTestData } from "tests/e2e/apply/fixtures/sf424d-fill-data";
-import { getOpportunityId } from "tests/e2e/get-opportunityId-utils";
+import { getOpportunityId } from "tests/e2e/utils/get-opportunityId-utils";
 import playwrightEnv from "tests/e2e/playwright-env";
 import { VALID_TAGS } from "tests/e2e/tags";
 import { authenticateE2eUser } from "tests/e2e/utils/authenticate-e2e-user-utils";
