+++ frontend/tests/e2e/apply/forms/happy-path/specs/happy-path-sfLLL.spec.ts
@@ -6,7 +6,7 @@
 import { test } from "@playwright/test";
 import { SFLLL_TEST_DATA } from "tests/e2e/apply/fixtures/sfLLL-field-definitions";
 import { SFLLL_FORM_CONFIG } from "tests/e2e/apply/fixtures/sfLLL-fill-data";
-import { getOpportunityId } from "tests/e2e/get-opportunityId-utils";
+import { getOpportunityId } from "tests/e2e/utils/get-opportunityId-utils";
 import playwrightEnv from "tests/e2e/playwright-env";
 import { VALID_TAGS } from "tests/e2e/tags";
 import { authenticateE2eUser } from "tests/e2e/utils/authenticate-e2e-user-utils";
