+++ frontend/tests/e2e/utils/get-opportunityId-utils.ts
@@ -1,8 +1,8 @@
 import {
   testdata_local_environment,
   testdata_staging_environment,
-} from "./opportunity-id-data";
-import playwrightEnv from "./playwright-env";
+} from "../opportunity-id-data";
+import playwrightEnv from "../playwright-env";
 
 export function getOpportunityId() {
   const { targetEnv } = playwrightEnv;
